Move money_field as its also part of the AR extension
