+++ lib/money_extensions.rb
@@ -4,12 +4,12 @@ module MoneyExtensions
   require 'money_extensions/integer'
   require 'money_extensions/split_between'
   require 'money_extensions/html_format'
-  require 'money_extensions/money_field'
   require 'money_extensions/numeric'
   require 'money_extensions/string'
 
   # TODO: Defer requiring the active record extension to who needs it
   if Module.const_defined?('ActiveRecord')
+    require 'money_extensions/active_record/money_field'
     require 'money_extensions/active_record/extensions'
   end
