+++ lib/money_extensions/active_record/money_field.rb
